+++ app/src/main/java/woowacourse/shopping/data/repository/CartRepository.kt
@@ -1,0 +1,11 @@
+package woowacourse.shopping.data.repository
+
+import woowacourse.shopping.model.Product
+
+interface CartRepository {
+    suspend fun addCartProduct(product: Product)
+
+    suspend fun getAllCartProducts(): List<Product>
+
+    suspend fun deleteCartProduct(id: Long)
+}
