+++ app/src/main/java/woowacourse/shopping/data/repository/DBCartRepository.kt
@@ -1,0 +1,22 @@
+package woowacourse.shopping.data.repository
+
+import woowacourse.shopping.data.db.CartProductDao
+import woowacourse.shopping.data.mapper.toEntity
+import woowacourse.shopping.data.mapper.toModel
+import woowacourse.shopping.model.Product
+
+class DBCartRepository(
+    private val cartProductDao: CartProductDao,
+) : CartRepository {
+    override suspend fun addCartProduct(product: Product) {
+        cartProductDao.insert(product.toEntity())
+    }
+
+    override suspend fun getAllCartProducts(): List<Product> {
+        return cartProductDao.getAll().map { it.toModel() }
+    }
+
+    override suspend fun deleteCartProduct(id: Long) {
+        cartProductDao.delete(id)
+    }
+}
