+++ app/src/main/java/woowacourse/shopping/data/repository/DefaultCartRepository.kt
@@ -1,0 +1,19 @@
+package woowacourse.shopping.data.repository
+
+import woowacourse.shopping.model.Product
+
+class DefaultCartRepository : CartRepository {
+    private val cartProducts: MutableList<Product> = mutableListOf()
+
+    override suspend fun addCartProduct(product: Product) {
+        cartProducts.add(product)
+    }
+
+    override suspend fun getAllCartProducts(): List<Product> {
+        return cartProducts.toList()
+    }
+
+    override suspend fun deleteCartProduct(id: Long) {
+        cartProducts.removeIf { it.id == id }
+    }
+}
